Add basic swapchain creation
A swapchain can now be created while optionally passing in the old
swapchain, useful for recreation. Destruction has been implemented with
the `Destroy` trait, as doing so with the `Drop` trait requires each
droppable object to store a reference to the Device. Passing around an
Arc<Mutex<Device>> can work, but choosing to destroy the objects
manually may prove easier in the long term.
Additionally, an initial framework for a delayed deletion queue has been
created, to handle 'zombie' objects (objects that are still in use by
the system, but are not safe to use anymore) such as the old swapchain.
This isn't being used yet, but will be used later to handle things like
per-frame resources.