Add clear-colour rendering, swapchain recreation, and other improvements
The idea originally was to have the application not require the Ash
crate at all, and instead have the rendering-engine crate provide all
needed functionality. I've decided instead to have rendering-engine be a
thin abstraction layer over Vulkan instead, with the application needing
to still use Ash heavily. This approach has the advantage of giving the
app nearly full control of Vulkan with less assumptions needed by the
abstraction.
Using this, the demo application now renders a clear colour to the
surface each frame with the new Renderer object. This encapsulates the
various objects required for rendering that are created by the app, and
handles drawing, swapchain recreation, and teardown. It also handles
Winit events to pause/unpause, trigger swapchain recreation, and draw
the frames.