Add shaders, pipelines, and draw a triangle
SPIR-V shaders can now be loaded using the Shader struct, which performs
reflection on the shader to determine the shader stage, input binding
description, and input attribute descriptions. Currently the vertices
are hardcoded so this doesn't do much, but once vertex buffers are
implemented it should handle them automatically and can be extended
later for push constant ranges and descriptor sets.
The PipelineBuilder struct has also been added to simplify the creation
of Pipeline objects, with many of the options configurable.
This has been used to draw a single triange hardcoded in the vertex
shader.