Implement vertex and index buffers to draw more complex geometry
This required a new Buffer struct that represents a Vulkan buffer, which
is allocated using the gpu_allocator crate. When creating a buffer with
data, the data is copied first to a staging buffer, before being copied
to a device-local buffer.
To support easy buffer copying, the Device struct has been extended with
'immediate submission' resources including a command pool/buffer to
allow immediately executing one-time functions when needed.
Vertices are no longer hardcoded in the vertex shader but just in the
renderer, though model loading can now be implemented instead.