Improve the material system and make use of emissive textures
Model loading now reads the materials from the glTF file, and a
MaterialData object is created for each. This object tracks the indexes
into the texture array for the various PBR textures (metallic/roughness,
normal, emissive etc).
These objects are stored in a storage buffer bound to a new descriptor
set, though currently the index into this buffer is hardcoded in the
fragment shader.
Only the base colour factor, base colour texture, and emissive textures
are implemented, though extending the shader to support the remaining
PBR textures should now be possible.