Modify model loading to create the buffers, and simplify drawing
The ModelData object now stores the vertex and optional index buffers,
creating them at the end of model loading.
The vertex and index offsets have been modified to store both the start
and count, avoiding the need to calculate the vertex/index count at draw
time.
The next step is to create a more convenient abstraction over the
creation of meshes and other assets, to support multiple models in a
scene.