Split model loading to a separate file
The ModelData struct has been renamed to Mesh, and load_model moved to
Mesh::new(). This struct also implements Destroy, and a new trait Draw,
as another step towards easily rendering multiple meshes at once.

The next step will be to add per-mesh transformations, and reworking
how vertex/transformation data is sent to the shader.

I'd like to support instancing soon as well, and work on an abstraction
for descriptor pool/set creation and management. This should tie in to
the work on bindless rendering.