Partial fix for #4 for the Instance object
`InstanceBuilder::enable_validation_layers()` has been removed, and
replaced with functions to specify the layers and extensions to be
enabled instead. This makes the distinction between instance-level
extensions required by the application and those required by the engine.

If a window handle is given, the `ash_window` crate is used to get the
required extensions and enable them automatically. It may be better to
not do this and have the application add the required extensions
instead, but I don't yet know of a situation where that would be useful.