Finish fix for #4 by improving extension requests for the Device
Fixes #4

The DeviceBuilder has been added to allow the application to specify
the extensions and features required. When building the device, these
extensions/features are enabled alongside those required by the engine.

Feature checking is not yet implemented as it seems to be impossible to
do automatically. Ash does not expose any helper functionality for this
either, so short of manually writing checks for every current extension
feature, the onus will be on the user to ensure their device supports
the required functionality.

To make the InstanceBuilder and DeviceBuilder consistent, the window
handle is no longer specified through a function on the InstanceBuilder,
but as an optional parameter when calling InstanceBuilder::build().
This matches passing the optional surface when calling the new
DeviceBuilder::builder().