Fix window resizing, and improve cursor grabbing
Window resizing has seemingly been broken since the depth image was
implemented, as it was never recreated to the new swapchain size. That
has now been fixed.

Additionally, the window event handling has been extended to improve
how cursor grabbing works. The cursor will not be grabbed when the app
starts, but clicking inside the window will grab it and enable camera
control until the escape key is pressed.