Begin work on moving the glTF loading into the engine
The glTF model loading is being moved in the scene module of the engine,
which will include a proper implementation for glTF that should be able
to handle all the necessary parts for vertex skinning and animations.

The in-progress Scene::load() is an improved version of Mesh::new(),
but having the scene structure represent the glTF structure with Nodes,
Meshes, and Primitives.

Calculating the transformation matrix for the node to take into account
the parent transforms is unimplemented at the moment, as getting the
correct order seems somewhat complex and will need more thought.