Add an initial implementation for animation
Textures are now stored in a slab (from the `slab` crate), in order to
help support multiple scenes in the future. This is tangential to the
implementation for animations, but has been included nonetheless.

In the `scene` module, structs have been added to represent glTF skins,
animation channels, and animations. These are read during scene loading
and populated with the necessary data.

The `load_node` function has been modified to fix an oversight, where
only the first level of children for root nodes would be loaded. This
only became a problem when loading certain models. Instead, the
`scene_nodes` vector is now pre-populated with default nodes, and each
node will add itself to the vector at it's respective index when loaded.
The parent-child loading has been simplified as loading all children
before the parent is not actually necessary.

The scene loading also creates a buffer to store the matrix for each
joint in the scene. This is updated with each call to
`Scene::update_joints()`, and it is the application's responsibility
to bind this buffer to the correct descriptor set to be used in the
vertex shader.

In the `renderer` module, the joint buffer is bound to the previously
created texture descriptor set at binding 1, to match the vertex shader.
In `Renderer::draw_frame()`, the time since the last draw is passed to
`Scene::update_animations()`.

The implementation is currently rather messy, especially on the
application side, however this will be the target for tidying up next by
creating a proper abstraction for descriptor sets, among other things.